Implement mesh part material selection and rendering

+++ engine/src/Resources/MaterialData.hpp
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <cstdint>
+
 namespace kokko
 {
 
 struct MaterialId
 {
-	unsigned int i;
+	uint16_t i;
 
 	bool operator==(MaterialId other) { return i == other.i; }
 	bool operator!=(MaterialId other) { return operator==(other) == false; }
